requirements.txt
numpy>=1.24
pandas>=1.5
streamlit>=1.27
plotly>=5.17
PyYAML>=6.0
highspy>=1.6
XlsxWriter>=3.1
openpyxl>=3.1
